Оптимизация работы фильтра и сортиовки
